rename naver to place in store
