+++ src/store/place/sagas/index.ts
@@ -2,6 +2,6 @@ import {searchPlaces} from './searchPlaces';
 import * as actions from "../actions"
 import { takeEvery } from 'redux-saga/effects';
 
-export function* naverSaga() {
+export function* placeSaga() {
   yield takeEvery(actions.SEARCH_PLACES, searchPlaces);
 }
