+++ src/store/place/sagas/searchPlaces/DUMMY_DATA.ts
