+++ src/store/place/sagas/searchPlaces/index.ts
